Adaptive updates
Moved the code from Adaptive into GUI, it no longer replicates effort.

Adaptive, now has a signal timeout, i.e. I kill the process if it goes over time.

Adaptive, I added messaging to it so it could display status as it runs

AHREA-logo.pgm - the updated logo with text added to it.